Hide Mezzanine header from pop
Mezzanine header belonging to the page inside the popup shows up
on top of the original page header, and also displaces the dialog
title bar. So, hide it with some css in the template. This way,
any page calling the popup will have the header hidden without
having to do anything on the original page.