Correct addEvent function
Use django's default addEvent implementation (in contrib/admin/static/admin/js/core.js), or if not provided then define our own based on django's implementation. This fixes a bug that occurs when there are multiple horizontal filtered many-to-many fields on an admin in IE10 (and perhaps other versions of IE).