Fix "The requested Folder does not exist" message.
When I go to edit a FileBrowseField on an existing model, this error
message is flashed and it goes to the root media directory rather than
whatever directory the file is in.

The reason for this is that this code replaces the directory with an
absolute path. I have no idea why. The fb_browse view clearly expects a
relative path:

    abs_path = os.path.join(get_directory(), path)